% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pull Agricultural Survey data using the USDA ARMS API}
\usage{
pull_data(year, report, state = "all", farmtype = 1, variable = "all")
}
\arguments{
\item{year}{A vector of years to filter data by}

\item{report}{A vector of reports to fetch}

\item{state}{A vector of states to filter data by}

\item{farmtype}{A vector of farmtypes to filter data by}

\item{variable}{A vector of variables to fetch}
}
\value{
A \code{data.frame} with survey data filtered according to 
input arguments
}
\description{
Pull Agricultural Survey data using the USDA ARMS API
}
