% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_metadata.R
\name{refresh_metadata}
\alias{refresh_metadata}
\title{Refresh metadata corresponding to the USDA ARMS API}
\usage{
refresh_metadata()
}
\value{
A \code{list} of metadata items
}
\description{
Refresh metadata corresponding to the USDA ARMS API
}
